+++ mobile/android/geckoview/src/main/java/org/mozilla/gecko/gfx/CompositorSurfaceManager.java
@@ -7,7 +7,6 @@ package org.mozilla.gecko.gfx;
 
 import android.os.RemoteException;
 import android.view.Surface;
-import android.view.SurfaceControl;
 import org.mozilla.gecko.annotation.WrapForJNI;
 
 public final class CompositorSurfaceManager {
@@ -20,9 +19,8 @@ public final class CompositorSurfaceManager {
   }
 
   @WrapForJNI(exceptionMode = "nsresult")
-  public synchronized void onSurfaceChanged(
-      final int widgetId, final Surface surface, final SurfaceControl surfaceControl)
+  public synchronized void onSurfaceChanged(final int widgetId, final Surface surface)
       throws RemoteException {
-    mManager.onSurfaceChanged(widgetId, surface, surfaceControl);
+    mManager.onSurfaceChanged(widgetId, surface);
   }
 }