+++ netwerk/protocol/http/nsHttp.h
@@ -407,7 +407,7 @@ static inline bool AllowedErrorForHTTPSRRFallback(nsresult aError) {
          aError == NS_ERROR_UNKNOWN_HOST || aError == NS_ERROR_NET_TIMEOUT;
 }
 
-bool SecurityErrorToBeHandledByTransaction(nsresult aReason);
+bool SecurityErrorThatMayNeedRestart(nsresult aReason);
 
 }  // namespace net
 }  // namespace mozilla