+++ taskcluster/gecko_taskgraph/util/attributes.py
@@ -75,50 +75,6 @@ _COPYABLE_ATTRIBUTES = (
 )
 
 
-def attrmatch(attributes, **kwargs):
-    """Determine whether the given set of task attributes matches.  The
-    conditions are given as keyword arguments, where each keyword names an
-    attribute.  The keyword value can be a literal, a set, or a callable.  A
-    literal must match the attribute exactly.  Given a set, the attribute value
-    must be in the set.  A callable is called with the attribute value.  If an
-    attribute is specified as a keyword argument but not present in the
-    attributes, the result is False."""
-    for kwkey, kwval in kwargs.items():
-        if kwkey not in attributes:
-            return False
-        attval = attributes[kwkey]
-        if isinstance(kwval, set):
-            if attval not in kwval:
-                return False
-        elif callable(kwval):
-            if not kwval(attval):
-                return False
-        elif kwval != attributes[kwkey]:
-            return False
-    return True
-
-
-def keymatch(attributes, target):
-    """Determine if any keys in attributes are a match to target, then return
-    a list of matching values. First exact matches will be checked. Failing
-    that, regex matches and finally a default key.
-    """
-    # exact match
-    if target in attributes:
-        return [attributes[target]]
-
-    # regular expression match
-    matches = [v for k, v in attributes.items() if re.match(k + "$", target)]
-    if matches:
-        return matches
-
-    # default
-    if "default" in attributes:
-        return [attributes["default"]]
-
-    return []
-
-
 def match_run_on_projects(project, run_on_projects):
     """Determine whether the given project is included in the `run-on-projects`
     parameter, applying expansions for things like "integration" mentioned in