+++ taskcluster/gecko_taskgraph/util/workertypes.py
@@ -4,9 +4,8 @@
 
 
 from mozbuild.util import memoize
-
-from .keyed_by import evaluate_keyed_by
-from .attributes import keymatch
+from taskgraph.util.attributes import keymatch
+from taskgraph.util.keyed_by import evaluate_keyed_by
 
 WORKER_TYPES = {
     "gce/gecko-1-b-linux": ("docker-worker", "linux"),